backend/requirements.txt
fastapi
uvicorn
cassandra-driver
python-dotenv
black
psycopg
psycopg2-binary
sqlalchemy
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
cors
python-dotenv
stripe
PyJWT
requests
